early dawn
The initial part of the Development plan is started, along with a full yoga program by Adriene. This plan as of now is enough to begin a personal development plan if done in a loop. I'll continue to edit and update, revising and adding days over time. This is in active development.